refactor: Extract options{consumer,provider} to file
