+++ packages/cozy-procedures/src/index.js
@@ -10,41 +10,7 @@ import _Duration from './containers/Duration'
 import _PersonalDataForm from './containers/PersonalDataForm'
 import _DocumentsDataForm from './components/Documents'
 
-import { PageLayout, PageFooter, PageContent } from 'cozy-ui/transpiled/react'
-
-const ProcedureOptionsContext = createContext({})
-
-const DEFAULT_OPTIONS = {
-  components: {
-    PageLayout,
-    PageFooter,
-    PageContent
-  }
-}
-
-const optionsProvider = (Component, procedureOptions = DEFAULT_OPTIONS) => {
-  const Wrapped = props => {
-    return (
-      <ProcedureOptionsContext.Provider value={procedureOptions}>
-        <Component {...props} />
-      </ProcedureOptionsContext.Provider>
-    )
-  }
-  Wrapped.displayName = `procedureOptionsProvider(${Component.displayName ||
-    Component.name})`
-  return Wrapped
-}
-
-const optionsConsumer = Component => props => {
-  const Wrapped = (
-    <ProcedureOptionsContext.Consumer>
-      {options => <Component {...options} {...props} />}
-    </ProcedureOptionsContext.Consumer>
-  )
-  Wrapped.displayName = `procedureOptionsConsumer(${Component.displayName ||
-    Component.name})`
-  return Wrapped
-}
+import { optionsConsumer, optionsProvider } from './ProcedureOptions'
 
 const Amount = optionsConsumer(_Amount)
 const Duration = optionsConsumer(_Duration)
